added modified date to model mixin
also required changes to directory mixin to accommodate changes to the
save process. That is, modified should match created if same transaction,
but created is set in save for directories on first go. Thus I needed to
be able to override the modified setting in the date mixin save function.