reworked modified_at in date mixin
I realized that the way I was approaching the modified at may not have
been correct. I've changed it such that now the modified_at is not set at
model creation, just on successive saves. So if a model has been saved
only once (never modified) then it will not have a modified date. If it
would be advantageous to have the previous behaivor (modified == created
on creation), revert this commit.